+++ docs/CHANGELOG.md
@@ -12,6 +12,12 @@ All notable changes to OrderScrapper iOS project will be documented in this file
 
 ## [Unreleased]
 
+## [0.1.8] - 2021-04-07
+
+### Added
+- Added the AnalyticsProvider to delegate the event logging mechanism to the app, in case the app has implemented 
+  any analytics. If there is no analytics implemented in the app, the SDK logs the event in Firebase as usual.
+
 ## [0.1.7] - 2021-03-31
 
 ### Fixed